# Remove unnecessary call to check if a path is ignored