--- a/lib/load-paths-handler.js
+++ b/lib/load-paths-handler.js
@@ -46,7 +46,7 @@ class PathLoader {
   }
 
   pathLoaded (loadedPath, done) {
-    if (!this.isIgnored(loadedPath) && !emittedPaths.has(loadedPath)) {
+    if (!emittedPaths.has(loadedPath)) {
       this.paths.push(loadedPath)
       emittedPaths.add(loadedPath)
     }