@rafeca 2019-02-27
> I see that the test suite is passing. That's a great sign. 😅 In addition to having a green test suite, can you describe the process you'll follow to verify that the change has not introduced any regressions? (This is [something we ask contributors to provide in performance-related pull requests in atom/atom](https://github.com/atom/atom/blame/2712b1f5055ccfff59b4497244107850553552cd/.github/PULL_REQUEST_TEMPLATE/performance_improvement.md#L31), but it doesn't look like we've incorporated that into atom/fuzzy-finder's pull request template yet.)

Thanks for the suggestion! I've just added some information on the PR description about the verification steps that I followed.